requirements.txt
# Core ML/NLP Libraries
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.2
accelerate>=0.24.0

# Vector Store
faiss-cpu>=1.7.4
# For GPU support, replace faiss-cpu with: faiss-gpu>=1.7.4

# Document Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.1.0

# LangChain for RAG utilities
langchain>=0.1.0
langchain-community>=0.0.10

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Logging & Monitoring
loguru>=0.7.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
ipython>=8.12.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
